I made a push to get authenication working using passport.  I originally used a basic stategy but realized that was incorrect thankfully the logic written for the basic strategy was essentially the same for the local strat. I also have gone ahead and finalized the procedure for adding new features to our code base. A new feature will go inside a folder named after that feature. Authenication for example will go inside the auth folder.
