+++ package.json
@@ -13,15 +13,18 @@
   },
   "dependencies": {
     "bcrypt": "^3.0.6",
+    "connect-pg-simple": "^6.0.1",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
     "dotenv": "^8.2.0",
     "express": "~4.16.1",
+    "express-session": "^1.17.0",
     "http-errors": "~1.6.3",
     "morgan": "~1.9.1",
     "passport": "^0.4.0",
     "passport-google-oauth20": "^2.0.0",
     "passport-http": "^0.3.0",
+    "passport-local": "^1.0.0",
     "pg": "^7.12.1",
     "pg-hstore": "^2.3.3",
     "pg-promise": "^9.3.3",
